dev: move DotWorkflowActions Component to UI library
