+++ core-web/libs/edit-content/src/lib/components/dot-edit-content-toolbar/dot-edit-content-toolbar.component.spec.ts
@@ -1,60 +1,20 @@
 import { createComponentFactory, Spectator } from '@ngneat/spectator/jest';
 
-import { SplitButton, SplitButtonModule } from 'primeng/splitbutton';
 import { ToolbarModule } from 'primeng/toolbar';
 
-import { DotCMSWorkflowAction } from '@dotcms/dotcms-models';
+import { DotWorkflowActionsComponent } from '@dotcms/ui';
 import { mockWorkflowsActions } from '@dotcms/utils-testing';
 
 import { DotEditContentToolbarComponent } from './dot-edit-content-toolbar.component';
 
-const WORKFLOW_ACTIONS_SEPARATOR_MOCK: DotCMSWorkflowAction = {
-    assignable: true,
-    commentable: true,
-    condition: '',
-    icon: 'workflowIcon',
-    id: '44d4d4cd-c812-49db-adb1-1030be73e69a',
-    name: 'SEPARATOR',
-    nextAssign: 'db0d2bca-5da5-4c18-b5d7-87f02ba58eb6',
-    nextStep: '43e16aac-5799-46d0-945c-83753af39426',
-    nextStepCurrentStep: false,
-    order: 0,
-    owner: null,
-    roleHierarchyForAssign: true,
-    schemeId: '85c1515c-c4f3-463c-bac2-860b8fcacc34',
-    showOn: ['UNLOCKED', 'LOCKED'],
-    metadata: {
-        subtype: 'SEPARATOR'
-    },
-    actionInputs: [
-        {
-            body: {},
-            id: 'assignable'
-        },
-        {
-            body: {},
-            id: 'commentable'
-        },
-        {
-            body: {},
-            id: 'pushPublish'
-        },
-        { body: {}, id: 'moveable' }
-    ]
-};
-
-const WORKFLOW_ACTIONS_MOCK = [
-    ...mockWorkflowsActions,
-    WORKFLOW_ACTIONS_SEPARATOR_MOCK,
-    ...mockWorkflowsActions
-];
+const WORKFLOW_ACTIONS_MOCK = [...mockWorkflowsActions, ...mockWorkflowsActions];
 
 describe('DotEditContentToolbarComponent', () => {
     let spectator: Spectator<DotEditContentToolbarComponent>;
 
     const createComponent = createComponentFactory({
         component: DotEditContentToolbarComponent,
-        imports: [ToolbarModule, SplitButtonModule],
+        imports: [ToolbarModule, DotWorkflowActionsComponent],
         detectChanges: false
     });
 
@@ -67,24 +27,19 @@ describe('DotEditContentToolbarComponent', () => {
         spectator.detectComponentChanges();
     });
 
-    it('should render an extra split button for each `SEPARATOR` Action', () => {
-        const splitButtons = spectator.queryAll(SplitButton);
-        expect(splitButtons.length).toBe(2);
+    it('should dot-workflow-actions component with the correct input', () => {
+        const component = spectator.query(DotWorkflowActionsComponent);
+        expect(component).toBeTruthy();
+        expect(component.actions).toEqual(WORKFLOW_ACTIONS_MOCK);
     });
 
-    it('should emit the action when click on a split button', () => {
+    it('should emit the action dot-workflow-actions emits the fired action', () => {
         const spy = jest.spyOn(spectator.component.actionFired, 'emit');
-        const splitButton = spectator.query('.p-splitbutton > button');
-        splitButton.dispatchEvent(new Event('click'));
+        const component = spectator.query(DotWorkflowActionsComponent);
 
+        component.actionFired.emit(WORKFLOW_ACTIONS_MOCK[0]);
+
+        expect(component).toBeTruthy();
         expect(spy).toHaveBeenCalledWith(WORKFLOW_ACTIONS_MOCK[0]);
-    });
-
-    it('should render a normal button is a group has only one action', () => {
-        spectator.setInput('actions', [WORKFLOW_ACTIONS_MOCK[0]]);
-        spectator.detectChanges();
-
-        const button = spectator.query('.p-button');
-        expect(button).not.toBeNull();
     });
 });
