+++ core-web/libs/ui/src/lib/components/dot-workflow-actions/dot-workflow-actions.component.html
@@ -1,0 +1,12 @@
+@for(action of groupedActions(); track $index; let first = $first) { @if(action.length > 1) {
+<p-splitButton
+    [styleClass]="!first ? 'p-button-outlined' : ''"
+    [model]="action | slice : 1"
+    [label]="action[0].label"
+    (onClick)="action[0].command()"></p-splitButton>
+} @else {
+<p-button
+    [styleClass]="!first ? 'p-button-outlined' : ''"
+    [label]="action[0].label"
+    (onClick)="action[0].command()"></p-button>
+} }
